getUserDataPath can be inlined to make it's dependencies more explicit
